Replace unchecked mul with checked mul in PCache
The purpose of this is to allow calculations which
would cause overflow to proceed, since if it is
overflowing, we'll never calculate primes that large
anyway and it won't matter that the calculated value
is wrong.

Still need to perform additional checks to make it
work more cleanly and to provide better guarentees.